refactor: LottoView 제거 및 LottoController 클래스로 병합
