+++ src/main/java/lotto/AppConfig.java
@@ -4,7 +4,6 @@ import lotto.controller.LottoController;
 import lotto.domain.ticket.generator.RandomTicketGenerator;
 import lotto.domain.ticket.generator.TicketGenerator;
 import lotto.service.LottoService;
-import lotto.view.LottoView;
 import lotto.view.input.InputView;
 import lotto.view.input.reader.ConsoleReader;
 import lotto.view.input.reader.Reader;
@@ -17,7 +16,6 @@ public class AppConfig {
     public final LottoController lottoController;
     public final LottoService lottoService;
     public final TicketGenerator ticketGenerator;
-    public final LottoView lottoView;
     public final OutputView outputView;
     public final InputView inputView;
     public final Reader reader;
@@ -26,10 +24,9 @@ public class AppConfig {
         this.reader = initReader();
         this.inputView = initInputView(reader);
         this.outputView = initOutputView();
-        this.lottoView = initLottoView(inputView, outputView);
         this.ticketGenerator = initTicketGenerator();
         this.lottoService = initLottoService(ticketGenerator);
-        this.lottoController = initLottoController(lottoService, lottoView);
+        this.lottoController = initLottoController(lottoService, inputView, outputView);
     }
 
     public static AppConfig getInstance() {
@@ -48,10 +45,6 @@ public class AppConfig {
         return new OutputView();
     }
 
-    private LottoView initLottoView(final InputView inputView, final OutputView outputView) {
-        return new LottoView(inputView, outputView);
-    }
-
     private TicketGenerator initTicketGenerator() {
         return new RandomTicketGenerator();
     }
@@ -60,8 +53,10 @@ public class AppConfig {
         return new LottoService(ticketGenerator);
     }
 
-    private LottoController initLottoController(final LottoService lottoService, final LottoView lottoView) {
-        return new LottoController(lottoService, lottoView);
+    private LottoController initLottoController(final LottoService lottoService,
+                                                final InputView inputView,
+                                                final OutputView outputView) {
+        return new LottoController(lottoService, inputView, outputView);
     }
 
 }
