+++ src/main/java/lotto/controller/LottoController.java
@@ -1,5 +1,12 @@
 package lotto.controller;
 
+import static lotto.view.output.OutputMessage.DIVIDING_LINE;
+import static lotto.view.output.OutputMessage.EMPTY_STRING;
+import static lotto.view.output.OutputMessage.REQUEST_BONUS_NUMBER;
+import static lotto.view.output.OutputMessage.REQUEST_MONEY;
+import static lotto.view.output.OutputMessage.REQUEST_WINNING_NUMBERS;
+import static lotto.view.output.OutputMessage.TITLE_OF_ANALYSIS;
+
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -10,16 +17,19 @@ import lotto.dto.AnalysisDto;
 import lotto.dto.TicketDto;
 import lotto.dto.WinningTicketDto;
 import lotto.service.LottoService;
-import lotto.view.LottoView;
+import lotto.view.input.InputView;
+import lotto.view.output.OutputView;
 
 public class LottoController {
 
     private final LottoService lottoService;
-    private final LottoView lottoView;
+    private final InputView inputView;
+    private final OutputView outputView;
 
-    public LottoController(final LottoService lottoService, final LottoView lottoView) {
+    public LottoController(final LottoService lottoService, final InputView inputView, final OutputView outputView) {
         this.lottoService = lottoService;
-        this.lottoView = lottoView;
+        this.inputView = inputView;
+        this.outputView = outputView;
     }
 
     public void run() {
@@ -31,7 +41,7 @@ public class LottoController {
     }
 
     public Tickets purchaseTickets() {
-        final int money = lottoView.requestMoney();
+        final int money = this.requestMoney();
         return lottoService.generateTickets(money);
     }
 
@@ -39,18 +49,51 @@ public class LottoController {
         final List<TicketDto> ticketDtos = tickets.getTickets().stream()
                 .map(TicketDto::toDto)
                 .collect(Collectors.toUnmodifiableList());
-        lottoView.announceTickets(ticketDtos);
+        this.announceTickets(ticketDtos);
     }
 
     public lotto.domain.ticket.Analysis calculateAnalysis(Tickets tickets) {
-        final WinningTicketDto winningTicketDto = lottoView.requestWinningTicket();
+        final WinningTicketDto winningTicketDto = this.requestWinningTicket();
         final WinningTicket winningTicket = winningTicketDto.toWinningTicket();
         return lottoService.generateAnalysis(tickets, winningTicket);
     }
 
     private void showAnalysis(final lotto.domain.ticket.Analysis analysis) {
         final AnalysisDto analysisDto = AnalysisDto.toDto(analysis);
-        lottoView.announceAnalysis(analysisDto);
+        this.announceAnalysis(analysisDto);
+    }
+
+    public int requestMoney() {
+        outputView.printMessage(REQUEST_MONEY);
+        return inputView.requestMoney();
+    }
+
+    public WinningTicketDto requestWinningTicket() {
+        final List<Integer> winningNumbers = requestWinningNumbers();
+        final int bonusNumber = requestBonusNumber();
+        return new WinningTicketDto(winningNumbers, bonusNumber);
+    }
+
+    private List<Integer> requestWinningNumbers() {
+        outputView.printMessage(REQUEST_WINNING_NUMBERS);
+        return inputView.requestWinningNumbers();
+    }
+
+    private int requestBonusNumber() {
+        outputView.printMessage(REQUEST_BONUS_NUMBER);
+        return inputView.requestBonusNumber();
+    }
+
+    public void announceTickets(List<TicketDto> ticketDtos) {
+        outputView.printTicketCount(ticketDtos);
+        outputView.printTickets(ticketDtos);
+    }
+
+    public void announceAnalysis(AnalysisDto analysis) {
+        outputView.printMessage(EMPTY_STRING);
+        outputView.printMessage(TITLE_OF_ANALYSIS);
+        outputView.printMessage(DIVIDING_LINE);
+        outputView.printAnalysis(analysis);
     }
 
 }
