--- src/main/java/lotto/view/LottoView.java
@@ -1,56 +1,0 @@
-package lotto.view;
-
-import static lotto.view.output.OutputMessage.*;
-
-import java.util.List;
-
-import lotto.dto.AnalysisDto;
-import lotto.dto.TicketDto;
-import lotto.dto.WinningTicketDto;
-import lotto.view.input.InputView;
-import lotto.view.output.OutputView;
-
-public class LottoView {
-
-    private final InputView inputView;
-    private final OutputView outputView;
-
-    public LottoView(InputView inputView, OutputView outputView) {
-        this.inputView = inputView;
-        this.outputView = outputView;
-    }
-
-    public int requestMoney() {
-        outputView.printMessage(REQUEST_MONEY);
-        return inputView.requestMoney();
-    }
-
-    public WinningTicketDto requestWinningTicket() {
-        final List<Integer> winningNumbers = requestWinningNumbers();
-        final int bonusNumber = requestBonusNumber();
-        return new WinningTicketDto(winningNumbers, bonusNumber);
-    }
-
-    private List<Integer> requestWinningNumbers() {
-        outputView.printMessage(REQUEST_WINNING_NUMBERS);
-        return inputView.requestWinningNumbers();
-    }
-
-    private int requestBonusNumber() {
-        outputView.printMessage(REQUEST_BONUS_NUMBER);
-        return inputView.requestBonusNumber();
-    }
-
-    public void announceTickets(List<TicketDto> ticketDtos) {
-        outputView.printTicketCount(ticketDtos);
-        outputView.printTickets(ticketDtos);
-    }
-
-    public void announceAnalysis(AnalysisDto analysis) {
-        outputView.printMessage(EMPTY_STRING);
-        outputView.printMessage(TITLE_OF_ANALYSIS);
-        outputView.printMessage(DIVIDING_LINE);
-        outputView.printAnalysis(analysis);
-    }
-
-}
